refactor(module-api): move RawVideoRelease to separate package
